feat: create main-service

+++ stats-service/server/src/main/java/ru/practicum/repository/AppRepository.java
@@ -5,6 +5,6 @@ import ru.practicum.model.App;
 
 import java.util.Optional;
 
-public interface AppRepository extends JpaRepository<App, Integer> {
+public interface AppRepository extends JpaRepository<App, Long> {
     Optional<App> findByName(String app);
 }
